requirements.txt
# It's recommanded to install torch and torchvision manually via https://pytorch.org/
# torch
# torchvision

opencv-python
numpy>=1.20.0
matplotlib>=3.3.0
pandas>=1.2.0
tqdm
tensorboard
tensorboardX
kornia
scikit-image>=0.26.0
pyyaml>=6.0.2
shapely>=2.1.2
seaborn>=0.13.2
pywebview>=6.1
ultralytics==8.4.21
fastapi>=0.128.0
uvicorn>=0.40.0
python-multipart
